Move api fixtures into tests/api
